Moves rollins to ss-based polling.
Poll using `ss` instead of the Web100 libraries.

This is fundamentally still a polling-based kludge. In the future, when we are
all running Linux 4.3+, this should be changed to use `ss -E`, which is
basically `tail -f` for connections on a server. Then we can be certain no
connections will be missed.